#7, test case failures.

+++ src/main/java/nl/vpro/jcr/criteria/advanced/impl/QueryExecutorHelper.java
@@ -21,7 +21,7 @@ package nl.vpro.jcr.criteria.advanced.impl;
 
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.function.IntSupplier;
+import java.util.function.LongSupplier;
 
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
@@ -65,7 +65,7 @@ public final class QueryExecutorHelper {
     public static AdvancedResultImpl execute(
         String stmt,
         String language,
-        IntSupplier queryCounter,
+        LongSupplier queryCounter,
         Session jcrSession,
         int maxResults,
         int offset,
@@ -114,7 +114,7 @@ public final class QueryExecutorHelper {
     public static AdvancedResultImpl execute(
         String stmt,
         String language,
-        IntSupplier queryCounter,
+        LongSupplier queryCounter,
         Session jcrSession,
         int maxResults,
         int offset,
